refactor: sidebar 스펙 분리
